# docs/swagger.yaml
basePath: /
definitions:
...
  handler.LoginVendorAdminRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handler.LoginVendorAdminResponse:
    properties:
      email:
        type: string
      name:
        type: string
      token:
        type: string
    type: object
...
  handler.RegisterVendorAdminRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      vendor_id:
        type: string
    required:
    - email
    - name
    - password
    - vendor_id
    type: object
...
host: localhost:8080
info:
  contact:
    email: Obie.kal22@gmail.com
    url: www.linkedin.com/in/obie-ananda-a87a64212
  description: API documentation for the FTGO PlashCash project.
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  termsOfService: http://example.com/terms/
  title: FTGO PlasCash Project
  version: "1.0"
paths:
...
  /vendor-admin/login:
    post:
      consumes:
      - application/json
      description: Authenticates a vendor admin using email and password, returning
        a JWT token.
      parameters:
      - description: Vendor Admin Login Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginVendorAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token
          schema:
            $ref: '#/definitions/handler.LoginVendorAdminResponse'
        "400":
          description: Invalid email or password
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login for vendor admin
      tags:
      - Vendor Admin
  /vendor-admin/recycle/{transaction_id}:
    post:
      consumes:
      - application/json
      description: Processes a customer recycling transaction, calculates rewards,
        and generates a token.
      parameters:
      - description: Transaction ID
        in: path
        name: transaction_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Recycling facilitated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid or unauthorized transaction
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Facilitate customer recycling
      tags:
      - Vendor Admin
  /vendor-admin/register:
    post:
      consumes:
      - application/json
      description: Registers a new vendor admin with name, email, password, and vendor
        ID.
      parameters:
      - description: Vendor Admin Registration Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterVendorAdminRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Vendor admin registered successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request or email already registered
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a vendor admin
      tags:
      - Vendor Admin
  /vendor-admin/transactions:
    get:
      consumes:
      - application/json
      description: Retrieves all transactions associated with the authenticated vendor.
      produces:
      - application/json
      responses:
        "200":
          description: Transactions fetched successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get transactions for a vendor
      tags:
      - Vendor Admin
  /vendor-admin/vending-machine/{vending_machine_id}/pickup:
    post:
      consumes:
      - application/json
      description: Creates a pickup request for a full vending machine.
      parameters:
      - description: Vending Machine ID
        in: path
        name: vending_machine_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pickup requested successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid vending machine ID or vending machine is not full
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Request a vending machine pickup
      tags:
      - Vendor Admin
  /vendor-admin/vending-machine/{vending_machine_id}/status:
    get:
      consumes:
      - application/json
      description: Retrieves the status of a specific vending machine.
      parameters:
      - description: Vending Machine ID
        in: path
        name: vending_machine_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Vending machine status retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid vending machine ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get vending machine status
      tags:
      - Vendor Admin
...
swagger: "2.0"